Moving history out of ui since it has nothing to do with ui
