+++ lib/gcli/ui/inputter.js
@@ -48,7 +48,7 @@ var KeyEvent = event.KeyEvent;
 
 var Status = require('gcli/types').Status;
 var Templater = require('gcli/ui/domtemplate').Templater;
-var History = require('gcli/ui/history').History;
+var History = require('gcli/history').History;
 
 var inputterCss = require('text!gcli/ui/inputter.css');
