+++ lib/gclitest/testHistory.js
@@ -7,7 +7,7 @@
 define(function(require, exports, module) {
 
 var t = require('test/assert');
-var History = require('gcli/ui/history').History;
+var History = require('gcli/history').History;
 
 exports.setup = function() {
 };
